More work on TokenList
