+++ storefront/components/Tokens/TokenFontSize/TokenFontSize.module.css
@@ -1,7 +1,7 @@
 .font {
-  height: 76px;
-  border-radius: 4px 4px 0 0;
-  border-bottom: 1px solid #dcdcdc;
+  height: 100%;
+  width: 100%;
+  border-radius: 4px;
   display: flex;
   align-items: center;
   padding-left: 24px;
