+++ storefront/components/Tokens/TokenShadow/TokenShadow.module.css
@@ -2,5 +2,4 @@
   height: 100%;
   width: 100%;
   border-radius: 4px;
-  border: 1px solid #dcdcdc;
 }
