+++ storefront/components/Tokens/TokenShadow/TokenShadow.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+import classes from './TokenShadow.module.css';
+
+interface TokenColorProps {
+  value: string;
+}
+
+const TokenShadow = ({ value }: TokenColorProps) => {
+  return (
+    <div
+      style={{ boxShadow: value }}
+      className={classes.color}
+    ></div>
+  );
+};
+
+export { TokenShadow };
