--- storefront/components/Tokens/Tokens.module.css
@@ -1,65 +1,0 @@
-.tokens {
-  box-shadow: none !important;
-}
-
-.cards {
-  display: grid;
-  grid-row-gap: var(--grid-gap);
-  grid-column-gap: var(--grid-gap);
-  grid-template-columns: repeat(auto-fit, minmax(min(100%, 320px), 1fr));
-  margin-top: var(--spacing-4);
-}
-
-.card {
-  background-color: #ffffff;
-  border-radius: 4px;
-  border: 1px solid #dcdcdc;
-}
-
-.color {
-  height: 76px;
-  border-radius: 4px 4px 0 0;
-  border-bottom: 1px solid #dcdcdc;
-}
-
-.copy {
-  background-color: #eeeeee;
-  margin: 16px;
-  padding: 11px 15px;
-  border-radius: 4px;
-  font-size: 16px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.copyIcon {
-  margin-bottom: -2px;
-}
-
-.value {
-  margin: -4px 16px 16px;
-}
-
-.test {
-  width: 100%;
-  height: 70px;
-  overflow: hidden;
-  position: relative;
-}
-
-.test::before {
-  --pattern-repeated: 20;
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 10000%;
-  aspect-ratio: 1/1;
-  background: repeating-conic-gradient(
-      transparent 0 90deg,
-      rgb(227, 227, 227) 0 180deg
-    )
-    0 0 / calc(2% / var(--pattern-repeated)) calc(2% / var(--pattern-repeated))
-    round;
-}
