+++ storefront/layouts/PageLayout/PageLayout.module.css
@@ -84,7 +84,7 @@
 }
 
 .content {
-  padding: 24px 32px;
+  padding: 16px 32px 24px;
 }
 
 .content p {
