+++ storefront/pages/grunnleggende/designelementer/skygger.mdx
@@ -1,6 +1,7 @@
 import Meta from '../../../components/Meta/Meta';
 import { PageLayout } from '../../../layouts';
 import { Figma } from 'mdx-embed';
+import { TokenList } from '../../../components/Tokens/TokenList';
 
 <Meta
   title='Skygger'
@@ -27,9 +28,9 @@ Skygger kan hjelpe svaksynte til å identifisere komponenter. Bruk av skygger og
 
 Vi har ulike styrker på skyggene, fra xsmall til xlarge. De ulike styrkene brukes for å antyde høyden til overflaten. Overflater i høyere høyder har større skygger, mens de på lavere høyder bør ha mindre skygger. Skygger skal skape et hierarki slik at det som ligger over eller under noe annet kommer tydeligere frem.
 
-<Figma
-  title='Figma Shadows'
-  url='file/eeznFkuH7UletQmhN174VR/Felles---Design-Tokens?node-id=4522%3A8679&t=hEFVDudlLerIyl0x-1'
+<TokenList
+  type='shadow'
+  token='fds-shadow'
 />
 
 ### Eksempel
