+++ storefront/pages/grunnleggende/designelementer/typografi.mdx
@@ -1,6 +1,5 @@
 import Meta from '../../../components/Meta/Meta';
 import { PageLayout } from '../../../layouts';
-import { Tokens } from '../../../components/Tokens/Tokens';
 
 <Meta
   title='Typografi'
